remove flux, remove stable swap store, make claim actions into mutations
