+++ Frontend-v1-Original/package.json
@@ -25,7 +25,6 @@
     "bignumber.js": "^9.0.1",
     "cors": "^2.8.5",
     "eslint-plugin-import": "^2.27.5",
-    "flux": "^4.0.3",
     "lottie-react": "^2.2.1",
     "moment": "^2.29.1",
     "next": "12.2",
@@ -39,7 +38,6 @@
   "devDependencies": {
     "@tanstack/eslint-plugin-query": "^4.29.4",
     "@types/cors": "^2.8.13",
-    "@types/flux": "^3.1.11",
     "@types/node": "^18.16.1",
     "@types/uuid": "^9.0.0",
     "autoprefixer": "^10.4.13",
