+++ Frontend-v1-Original/stores/constants/actions.ts
@@ -1,4 +1,4 @@
-// GENERAL
+// TX_STATUS is used to update the status of a transaction in the store
 export const ERROR = "ERROR";
 export const STORE_UPDATED = "STORE_UPDATED";
 export const TX_PENDING = "TX_PENDING";
@@ -9,139 +9,3 @@ export const TX_STATUS = "TX_STATUS";
 export const TX_ADDED = "TX_ADDED";
 export const TX_OPEN = "TX_OPEN";
 export const WARNING = "WARNING";
-
-// export const CONNECTION_CONNECTED = "CONNECTION_CONNECTED";
-// export const CONNECTION_DISCONNECTED = "CONNECTION_DISCONNECTED";
-
-// export const CONFIGURE = "CONFIGURE";
-// export const CONFIGURE_RETURNED = "CONFIGURE_RETURNED";
-
-// export const ACCOUNT_CONFIGURED = "ACCOUNT_CONFIGURED";
-// export const ACCOUNT_CHANGED = "ACCOUNT_CHANGED";
-
-// export const GET_GAS_PRICES = "GET_GAS_PRICES";
-// export const GAS_PRICES_RETURNED = "GAS_PRICES_RETURNED";
-
-// export const UPDATED = "UPDATED";
-
-// export const CONFIGURE_SS = "CONFIGURE_SS";
-
-// export const GET_BALANCES = "GET_BALANCES";
-// export const BALANCES_RETURNED = "BALANCES_RETURNED";
-
-// export const SEARCH_ASSET = "SEARCH_ASSET";
-// export const ASSET_SEARCHED = "ASSET_SEARCHED";
-
-// export const BASE_ASSETS_UPDATED = "BASE_ASSETS_UPDATED";
-// export const SWAP_ASSETS_UPDATED = "SWAP_ASSETS_UPDATED";
-
-// LIQUIDITY
-// export const CREATE_PAIR_AND_STAKE = "CREATE_PAIR_AND_STAKE";
-// export const CREATE_PAIR_AND_DEPOSIT = "CREATE_PAIR_AND_DEPOSIT";
-// export const PAIR_CREATED = "PAIR_CREATED";
-
-// export const GET_CREATE_PAIR_BALANCES = "GET_CREATE_PAIR_BALANCES";
-// export const GET_CREATE_PAIR_BALANCES_RETURNED =
-//   "GET_CREATE_PAIR_BALANCES_RETURNED";
-
-// export const ADD_LIQUIDITY = "ADD_LIQUIDITY";
-// export const LIQUIDITY_ADDED = "LIQUIDITY_ADDED";
-
-// export const STAKE_LIQUIDITY = "STAKE_LIQUIDITY";
-// export const LIQUIDITY_STAKED = "LIQUIDITY_STAKED";
-
-// export const ADD_LIQUIDITY_AND_STAKE = "ADD_LIQUIDITY_AND_STAKE";
-// export const ADD_LIQUIDITY_AND_STAKED = "ADD_LIQUIDITY_AND_STAKED";
-
-// export const QUOTE_ADD_LIQUIDITY = "QUOTE_ADD_LIQUIDITY";
-// export const QUOTE_ADD_LIQUIDITY_RETURNED = "QUOTE_ADD_LIQUIDITY_RETURNED";
-
-// export const GET_LIQUIDITY_BALANCES = "GET_LIQUIDITY_BALANCES";
-// export const GET_LIQUIDITY_BALANCES_RETURNED =
-//   "GET_LIQUIDITY_BALANCES_RETURNED";
-
-// export const REMOVE_LIQUIDITY = "REMOVE_LIQUIDITY";
-// export const LIQUIDITY_REMOVED = "LIQUIDITY_REMOVED";
-
-// export const UNSTAKE_LIQUIDITY = "UNSTAKE_LIQUIDITY";
-// export const LIQUIDITY_UNSTAKED = "LIQUIDITY_UNSTAKED";
-
-// export const UNSTAKE_AND_REMOVE_LIQUIDITY = "UNSTAKE_AND_REMOVE_LIQUIDITY";
-// export const REMOVE_LIQUIDITY_AND_UNSTAKED = "REMOVE_LIQUIDITY_AND_UNSTAKED";
-
-// export const QUOTE_REMOVE_LIQUIDITY = "QUOTE_REMOVE_LIQUIDITY";
-// export const QUOTE_REMOVE_LIQUIDITY_RETURNED =
-//   "QUOTE_REMOVE_LIQUIDITY_RETURNED";
-
-// export const CREATE_GAUGE = "CREATE_GAUGE";
-// export const CREATE_GAUGE_RETURNED = "CREATE_GAUGE_RETURNED";
-
-// //SWAPS
-// export const SWAP = "SWAP";
-// export const SWAP_RETURNED = "SWAP_RETURNED";
-
-// export const QUOTE_SWAP = "QUOTE_SWAP";
-// export const QUOTE_SWAP_RETURNED = "QUOTE_SWAP_RETURNED";
-
-// export const WRAP_UNWRAP = "WRAP_UNWRAP";
-// export const WRAP_UNWRAP_RETURNED = "WRAP_UNWRAP_RETURNED";
-
-// VESTING
-// export const GET_VEST_NFTS = "GET_VEST_NFTS";
-// export const VEST_NFTS_RETURNED = "VEST_NFTS_RETURNED";
-
-// export const CREATE_VEST = "CREATE_VEST";
-// export const CREATE_VEST_RETURNED = "CREATE_VEST_RETURNED";
-
-// export const INCREASE_VEST_AMOUNT = "INCREASE_VEST_AMOUNT";
-// export const INCREASE_VEST_AMOUNT_RETURNED = "INCREASE_VEST_AMOUNT_RETURNED";
-
-// export const INCREASE_VEST_DURATION = "INCREASE_VEST_DURATION";
-// export const INCREASE_VEST_DURATION_RETURNED =
-//   "INCREASE_VEST_DURATION_RETURNED";
-
-// export const RESET_VEST = "RESET_VEST";
-// export const RESET_VEST_RETURNED = "RESET_VEST_RETURNED";
-
-// export const WITHDRAW_VEST = "WITHDRAW_VEST";
-// export const WITHDRAW_VEST_RETURNED = "WITHDRAW_VEST_RETURNED";
-
-// export const MERGE_NFT = "MERGE_NFT";
-// export const MERGE_NFT_RETURNED = "MERGE_NFT_RETURNED";
-
-//VOTING
-// export const VOTE = "VOTE";
-// export const VOTE_RETURNED = "VOTE_RETURNED";
-
-// export const GET_VEST_VOTES = "GET_VEST_VOTES";
-// export const VEST_VOTES_RETURNED = "VEST_VOTES_RETURNED";
-
-// export const CREATE_BRIBE = "CREATE_BRIBE";
-// export const BRIBE_CREATED = "BRIBE_CREATED";
-
-// export const GET_VEST_BALANCES = "GET_VEST_BALANCES";
-// export const VEST_BALANCES_RETURNED = "VEST_BALANCES_RETURNED";
-
-//REWARDS
-export const CLAIM_BRIBE = "CLAIM_BRIBE";
-export const CLAIM_X_BRIBE = "CLAIM_X_BRIBE";
-export const CLAIM_XX_BRIBE = "CLAIM_XX_BRIBE";
-
-export const CLAIM_REWARD = "CLAIM_REWARD";
-export const CLAIM_BLOTR_REWARD = "CLAIM_BLOTR_REWARD";
-export const CLAIM_REWARD_RETURNED = "CLAIM_REWARD_RETURNED";
-
-export const CLAIM_ALL_REWARDS = "CLAIM_ALL_REWARDS";
-export const CLAIM_ALL_REWARDS_RETURNED = "CLAIM_ALL_REWARDS_RETURNED";
-
-export const CLAIM_VE_DIST = "CLAIM_VE_DIST";
-export const CLAIM_VE_DIST_RETURNED = "CLAIM_VE_DIST_RETURNED";
-
-//WHITELIST
-// export const SEARCH_WHITELIST = "SEARCH_WHITELIST";
-// export const SEARCH_WHITELIST_RETURNED = "SEARCH_WHITELIST_RETURNED";
-// export const WHITELIST_TOKEN = "WHITELIST_TOKEN";
-// export const WHITELIST_TOKEN_RETURNED = "WHITELIST_TOKEN_RETURNED";
-
-//AUTO BRIBE
-// export const BRIBE_AUTO_BRIBE = "BRIBE_AUTO_BRIBE";
