+++ Frontend-v1-Original/stores/constants/contracts.ts
@@ -65,4 +65,6 @@ export const STABLE_TOKEN_ADDRESS =
 
 export const MSIG_ADDRESS = "0x13eeB8EdfF60BbCcB24Ec7Dd5668aa246525Dc51";
 
+export const OPTION_TOKEN = "0x9f9A1Aa08910867F38359F4287865c4A1162C202";
+
 //LFG
