+++ Frontend-v1-Original/stores/constants/contractsGoerli.ts
@@ -65,3 +65,5 @@ export const STABLE_TOKEN_ADDRESS =
   "0x80b5a32E4F032B2a058b4F29EC95EEfEEB87aDcd"; // need this
 
 export const MSIG_ADDRESS = "0x13eeB8EdfF60BbCcB24Ec7Dd5668aa246525Dc51";
+
+export const OPTION_TOKEN = "0x9f9A1Aa08910867F38359F4287865c4A1162C202";
